Refactor operator test imports
