+++ test/second_quantization/operators/fermionic/test_fermionic_op.py
@@ -23,9 +23,9 @@ from ddt import data, ddt, unpack
 from scipy.sparse import csc_matrix
 from scipy.sparse.linalg import eigs
 
-from qiskit_nature.operators.second_quantization import FermionicOp
+from qiskit_nature.second_quantization.operators import FermionicOp
 
-from .utils import str2list, str2str, str2tuple
+from ..utils import str2list, str2str, str2tuple
 
 
 @lru_cache(3)
