+++ test/second_quantization/operators/spin/test_spin_op.py
@@ -22,9 +22,9 @@ import numpy as np
 from ddt import data, ddt, unpack
 from qiskit.quantum_info import Pauli
 
-from qiskit_nature.operators.second_quantization import SpinOp
+from qiskit_nature.second_quantization.operators import SpinOp
 
-from .utils import str2list, str2str, str2tuple
+from ..utils import str2list, str2str, str2tuple
 
 
 @lru_cache(3)
